Add end-to-end tests in a separate test target

+++ WeatherApp/Weather Forecast Feature/AirQuality.swift
@@ -8,18 +8,18 @@
 import Foundation
 
 public struct AirQuality {
-    public let carbonMonoxide: Float
-    public let ozon: Float
-    public let nitrogenDioxide: Float
-    public let sulphurDioxide: Float
-    public let pm2_5: Float
-    public let pm10: Float
+    public let carbonMonoxide: Double
+    public let ozon: Double
+    public let nitrogenDioxide: Double
+    public let sulphurDioxide: Double
+    public let pm2_5: Double
+    public let pm10: Double
     public let usEpaIndex: Int  // 1 - Good, 2 - Moderate, 3 - Unhealthy for sensitive group,
                          // 4 - Unhealty, 5 - Very Unhealty, 6 - Hazardous
 
     public let gbDefraIndex: Int
     
-    public init(carbonMonoxide: Float, ozon: Float, nitrogenDioxide: Float, sulphurDioxide: Float, pm2_5: Float, pm10: Float, usEpaIndex: Int, gbDefraIndex: Int) {
+    public init(carbonMonoxide: Double, ozon: Double, nitrogenDioxide: Double, sulphurDioxide: Double, pm2_5: Double, pm10: Double, usEpaIndex: Int, gbDefraIndex: Int) {
         self.carbonMonoxide = carbonMonoxide
         self.ozon = ozon
         self.nitrogenDioxide = nitrogenDioxide
